+++ lib/helpers/index.ts
@@ -2,9 +2,13 @@
 COMMON HELPERS
 */
 export {List, Field} from './commons';
+
 /*
 STAGE HELPERS
 */
 export {GroupByPayload} from "./bucket";
+export {GroupByAutoPayload} from "./bucket-auto";
 export {ConditionPayload, EqualityPayload} from "./lookup";
 export {IgnorePayload, OnlyPayload} from "./project";
+export {SizePayload} from "./sample";
+export {CollectionPayload} from "./union-with";