+++ lib/helpers/union-with/index.test.ts
@@ -1,0 +1,15 @@
+import {CollectionPayload} from "./index";
+
+describe('unionWith helpers', () => {
+    test.each([
+        [CollectionPayload('cars'), { coll: 'cars', pipeline: [] }],
+        [CollectionPayload('collection', [
+            { $match: { test: 'unit' } }
+        ]), { coll: 'collection', pipeline: [{ $match: { test: 'unit' } }] }],
+    ])('%o should return %o', (
+        operation: any,
+        expected: any
+    ) => {
+        expect(operation).toEqual(expected);
+    });
+});